Update DB tests to use local Docker MySQL

+++ src/test/java/edu/cornell/library/integration/processing/GenerateSolrFieldsTest.java
@@ -25,19 +25,35 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import edu.cornell.library.integration.db_test.AbstractContainerBaseTest;
+import edu.cornell.library.integration.db_test.DbBaseTest;
 import edu.cornell.library.integration.marc.MarcRecord;
 import edu.cornell.library.integration.utilities.Config;
 import edu.cornell.library.integration.utilities.Generator;
 
-public class GenerateSolrFieldsTest {
-
-	static Config config;
+public class GenerateSolrFieldsTest extends DbBaseTest {
 	static GenerateSolrFields gen;
 
+//	@BeforeClass
+//	public static void setup() throws SQLException {
+//		List<String> requiredArgs = Config.getRequiredArgsForDB("Headings");
+//		requiredArgs.add("catalogClass");
+//		config = Config.loadConfig(requiredArgs);
+//		gen = new GenerateSolrFields(
+//				EnumSet.allOf(Generator.class),
+//				EnumSet.of(Generator.AUTHORTITLE,Generator.RECORDTYPE,Generator.CALLNUMBER,
+//						Generator.LANGUAGE, Generator.MARC, Generator.URL), "solrGenTest" );
+//		gen.setUpDatabase(config);
+//	}
+
 	@BeforeClass
-	public static void setup() throws SQLException {
+	public static void setup() throws IOException, SQLException {
+		useTestContainers = System.getenv(USE_TEST_CONTAINERS);
 		List<String> requiredArgs = Config.getRequiredArgsForDB("Headings");
 		requiredArgs.add("catalogClass");
+		if (useTestContainers != null) {
+			AbstractContainerBaseTest.setProperties();
+		}
 		config = Config.loadConfig(requiredArgs);
 		gen = new GenerateSolrFields(
 				EnumSet.allOf(Generator.class),
@@ -48,10 +64,12 @@ public class GenerateSolrFieldsTest {
 
 	@AfterClass
 	public static void cleanup() throws SQLException {
-		try (   Connection inventory = config.getDatabaseConnection("Current");
-				Statement stmt = inventory.createStatement()) {
-			stmt.executeUpdate("DROP TABLE solrGenTestData");
-			stmt.executeUpdate("DROP TABLE solrGenTestGenerators");
+		if (useTestContainers == null) {
+			try (	Connection inventory = config.getDatabaseConnection("Current");
+					Statement stmt = inventory.createStatement()) {
+				stmt.executeUpdate("DROP TABLE solrGenTestData");
+				stmt.executeUpdate("DROP TABLE solrGenTestGenerators");
+			}
 		}
 	}
